bug fixed

+++ utils/bloomfilter.h
@@ -11,7 +11,7 @@ private:
 	const static int kMaxSize = 1000000;
 	const static int kFuncSize = 5;
 	const string SaveFilePath = "./BloomFilter.txt";
-	const vector<unsigned long long>SeedSet = {131, 1919, 10007, 100000007, 1000000007};
+	const vector<uint64_t>SeedSet = {131, 1919, 10007, 100000007, 1000000007};
 	ifstream fin;
 	ofstream fout;
 	bool Slot_[kMaxSize];
